+++ app/system/modules/user/src/Controller/AuthController.php
@@ -16,19 +16,13 @@ class AuthController
      */
     public function loginAction($redirect = '')
     {
-
-		$module = App::module('system/user');
-		$url = App::url($module->config['login_redirect']);
-		$redirect = $redirect ? : $url;
+        if (!$redirect) {
+            $redirect = App::url(App::config('system/user')['login_redirect']);
+        }
 
 		if (App::user()->isAuthenticated()) {
 			return App::redirect($redirect);
 		}
-
-        return self::loginView([
-            'last_username' => App::session()->get(Auth::LAST_USERNAME),
-            'redirect' => $redirect
-        ]);
 
         return [
             '$view' => [
@@ -36,7 +30,8 @@ class AuthController
                 'name' => 'system/user/login.php'
             ],
             'last_username' => App::session()->get(Auth::LAST_USERNAME),
-            'redirect' => $redirect
+            'redirect' => $redirect,
+            'message' => __('Sign in to your account')
         ];
     }
 
@@ -94,67 +89,5 @@ class AuthController
             App::message()->error($error);
             return App::redirect((preg_replace('#(https?:)?//[^/]+#', '', App::url()->previous())));
         }
-    }
-
-    /**
-     * Returns a renderable view configuration for a login form
-     *
-     * @param array $args Login form arguments
-     *          $args['title'] : The page title
-     *          $args['last_username'] : Username to pre-fill in login form
-     *          $args['redirect'] : Path to redirect after successful login
-     *          $args['message'] : Overwrites default login box headline
-     * @return array View config that can be rendered, i.e. by returning from a controller action
-     */
-    public static function loginView($args = [])
-    {
-        $args = array_merge([
-            'title' => __('Login'),
-            'last_username' => '',
-            'redirect' => '',
-            'message' => __('Sign in to your account')
-        ], $args);
-
-        return [
-            '$view' => [
-                'title' => $args['title'],
-                'name' => 'system/user/login.php'
-            ],
-            'last_username' => $args['last_username'],
-            'redirect' => $args['redirect'],
-            'message' => $args['message']
-        ];
-    }
-
-    /**
-     * Returns a renderable view configuration for a generic message view
-     * @param array $args
-     *          $args['title'] : The page title
-     *          $args['message'] : The message to display
-     *          $args['success'] : Is this a success message [default: true]
-     *          $args['link'] : Optional button link to display [default: '']
-     *          $args['label'] : Optional button label ['default: '']
-     * @return array
-     */
-    public static function messageView($args = [])
-    {
-        $args = array_merge([
-            'title' => __('Account activation'),
-            'message' => '',
-            'success' => true,
-            'link' => '',
-            'label' => ''
-        ], $args);
-
-        return [
-            '$view' => [
-                'title' => $args['title'],
-                'name' => 'system/user/message.php'
-            ],
-            'message' => $args['message'],
-            'success' => $args['success'],
-            'link'    => $args['link'],
-            'label'   => $args['label'],
-        ];
     }
 }