+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 
 ## WIP
-- Fixed redirect after login
 
 ### Added
 - OpenGraph image option for site nodes
+
+#### Fixed
+- Fixed redirect after login
 
 ### Fixed
 - Fixed save shortcut in Firefox